client: make close async
There are two issues that we solve here:

1) The old close method was broken in that when we called close with a
channel that was open but not active, we would never call `closeNow`
which could definitely lead to FD leaks
2) We were seeing a lot of timeouts in shutdown because we are
submitting tasks synchrounously to an executor and so those threads get
tied up for up to 30s. Additionally, because the excecutor only has like
18 threads in it we end up blocking the submitting thread as well, which
causes us not to respect our overall timeout, which means shutdown can
take up to #connections * 30s which for an average instance is 5 hours.

This PR makes close async like _literally_ every other operation we do.
This should free up the submitting thread, and allow us to time out more
accurately.