fix connect race condition
There was a possible race here between a client sending a command and
the channel pipeline actually being configured after connect that could
cause the first command to simply get dropped. This prevents this from
happening by returning a `Future<ImapClient` in connect and having the
client factory return this future (as opposed to returning the client
directly). As part of this `create` has also been made private to
prevent users from getting the client without handling the future.